FEAT: Refactor component imports to new directory structure under 'vis'
