+++ src/pages/vis/battery.vue
