+++ src/pages/vis/diagrams.vue
