+++ src/pages/vis/iobroker-info.vue
