+++ src/pages/vis/light.vue
