+++ src/pages/vis/rooms.vue
