+++ src/router/routes.ts
@@ -30,22 +30,22 @@ export const routes = {
 export const routing = [
   {
     path: "/",
-    component: async () => await import("@/pages/page.vue"),
+    component: async () => await import("@/pages/vis/page.vue"),
     children: [
-      { component: async () => await import("@/pages/home.vue"), ...routes.home },
-      { component: async () => await import("@/pages/navigation.vue"), ...routes.navigation },
-      { component: async () => await import("@/pages/alexa.vue"), ...routes.alexa },
-      { component: async () => await import("@/pages/battery.vue"), ...routes.battery },
-      { component: async () => await import("@/pages/diagrams.vue"), ...routes.diagrams },
-      { component: async () => await import("@/pages/lists.vue"), ...routes.lists },
-      { component: async () => await import("@/pages/rooms.vue"), ...routes.window },
-      { component: async () => await import("@/pages/heating.vue"), ...routes.heating },
-      { component: async () => await import("@/pages/heating-control.vue"), ...routes.heatingControl },
-      { component: async () => await import("@/pages/iobroker-info.vue"), ...routes.iobrokerInfo },
-      { component: async () => await import("@/pages/calendar.vue"), ...routes.calender },
-      { component: async () => await import("@/pages/light.vue"), ...routes.light },
-      { component: async () => await import("@/pages/logs.vue"), ...routes.logs },
-      { component: async () => await import("@/pages/pv.vue"), ...routes.pv },
+      { component: async () => await import("@/pages/vis/home.vue"), ...routes.home },
+      { component: async () => await import("@/pages/vis/navigation.vue"), ...routes.navigation },
+      { component: async () => await import("@/pages/vis/alexa.vue"), ...routes.alexa },
+      { component: async () => await import("@/pages/vis/battery.vue"), ...routes.battery },
+      { component: async () => await import("@/pages/vis/diagrams.vue"), ...routes.diagrams },
+      { component: async () => await import("@/pages/vis/lists.vue"), ...routes.lists },
+      { component: async () => await import("@/pages/vis/rooms.vue"), ...routes.window },
+      { component: async () => await import("@/pages/vis/heating.vue"), ...routes.heating },
+      { component: async () => await import("@/pages/vis/heating-control.vue"), ...routes.heatingControl },
+      { component: async () => await import("@/pages/vis/iobroker-info.vue"), ...routes.iobrokerInfo },
+      { component: async () => await import("@/pages/vis/calendar.vue"), ...routes.calender },
+      { component: async () => await import("@/pages/vis/light.vue"), ...routes.light },
+      { component: async () => await import("@/pages/vis/logs.vue"), ...routes.logs },
+      { component: async () => await import("@/pages/vis/pv.vue"), ...routes.pv },
     ],
   },
   {
@@ -71,6 +71,6 @@ export const routing = [
   { component: async () => await import("@/pages/finance/description.vue"), ...routes.description },
   {
     path: "/:pathMatch(.*)*",
-    component: async () => await import("@/pages/error-page-not-found.vue"),
+    component: async () => await import("@/pages/vis/error-page-not-found.vue"),
   },
 ];
